feat: created separated component to handle main menu
